Fix to underflow exception of type YYMMDD
SAS files having attributes defined of type YYMMDD10 cause an underflow
exception in the parsers as it contains only 4 bytes in the array, but
a getDouble requires 8. This a quick fix by filling up the byte array
with zeros. I suspect that this issue needs to be handled further
upstream in the callstack.